+++ src/app/core/ase-header/ase-header.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ase-header',
+  standalone: true,
+  imports: [],
+  templateUrl: './ase-header.component.html',
+  styleUrl: './ase-header.component.scss'
+})
+export class AseHeaderComponent {
+
+}