+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add project basic configuration
 - Add basic layout
+- Add basic header